+++ vcf2baits.rb
@@ -58,7 +58,15 @@ def vcf2baits
 		for samp in @samples
 			tmptax.push($options.taxa[samp])
 		end
-		$options.taxa = tmptax
+		$options.taxa = tmptax.dup
+		unless $options.popcategories.nil? # Build popcategories hash for snp selection
+			tmpcat = $options.popcategories
+			$options.popcategories = {}
+			tmptax = tmptax.uniq # Using uniq! on a uniqued array returns nil
+			for taxon in 0 ... tmptax.size
+				$options.popcategories[tmptax[taxon]] = tmpcat[taxon]
+			end
+		end
 		write_file(".log.txt", "Variant Categories\nChromosome\tVariant\tCategory") if $options.log
 		@snps.keys.each { |chromo| @snps[chromo].each { |snp| snp.categorize } }
 		write_file(".log.txt", "") if $options.log # Add break line to make log easier to read